nps-call-decouple: refactor wip, moving logic to individual classes
