+++ app/uk/gov/hmrc/tai/model/domain/Employments.scala
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2020 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.tai.model.domain
+
+import uk.gov.hmrc.tai.model.tai.TaxYear
+
+//TODO would this be a singleton?
+//TODO do we need a separate class for the auditing
+class Employments {
+
+  def employmentsWithAccountsForYear(employments: Seq[Employment], year: TaxYear): Seq[Employment] =
+    employments.collect {
+      case employment if employment.hasAnnualAccountsForYear(year) =>
+        employment.copy(annualAccounts = employment.annualAccountsForYear(year))
+    }
+}
