+++ test/uk/gov/hmrc/tai/model/domain/EmploymentsSpec.scala
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2020 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.tai.model.domain
+
+import org.scalatestplus.play.PlaySpec
+import uk.gov.hmrc.domain.Generator
+import uk.gov.hmrc.http.HeaderCarrier
+
+import scala.util.Random
+
+class EmploymentsSpec extends PlaySpec {
+
+  implicit val hc: HeaderCarrier = HeaderCarrier()
+
+  trait EmploymentsSetup {
+    val nino = new Generator(new Random).nextNino
+    val testEmployments = new Employments
+  }
+  //TODO tests
+
+}
